Fix DraftJS error when pasting over non-collapsed selection
addText uses insertText, so the handlePastedText method was failing if a user
was highlighting text and then pasting to replace the highlighted text.

(error was: Target range must be collapsed for `insertText`)

A similar method, replaceText, has been aded, and the handlePastedText method
has been updated to use replaceText instead.