Fix blockquote exiting behaviour
Before this change, exiting a blockquote required users to press
`SHIFT+ENTER`. That was unexpected behaviour, as in pure markdown
blockquotes stop at a new line.

After this change, blockquotes behave like headings: When `ENTER` is
pressed a new empty block is inserted. This is more expected, and to
continue making blockquotes one can continue inserting `>` at the
beginning of lines.