Use standalone draft-js-prism-plugin
The Prism plugin @ngs wrote in this repo works great, I wanted to use it
too so I took it
and made it it's own standalone package:
[`draft-js-prism-plugin`](https://github.com/withspectrum/draft-js-prism-plugin)

This PR just deletes your implementation and replaces it with an import
from `draft-js-prism-plugin`. Since the standalone package is basically
just a copy & paste this works just as well as before.