Fix deleting styled blocks
When the first thing in an editor is an styled block you cannot get rid
of it, it'll always stay styled, because backspace only deletes the
content.

This patch fixes the issue by introducing custom behaviour for this
specific condition: If we're in the first block of the editor, it's a
styled block, the block is empty and we press backspace we change the
block type to 'unstyled'.

Fixes https://github.com/ngs/draft-js-markdown-shortcuts-plugin/issues/31