More lenient code block regex
This issue popped up in combination with the `draft-js-image-plugin`.
When an image is inserted and then deleted, the image plugin leaves an
empty line after the image.

Because this empty line is in the same block it breaks the code block
regex because suddenly it's "``` " (notice the space) rather than
"```", so one couldn't insert a code block right before a deleted image
anymore.

By changing the code block regex to be ever so slightly more lenient
with whitespace we fix the issue.