Unsticky inline styles
This "unstickies" inline styles, so that when you type *bold* text,
backspace to the ending edge of the bold text and start typing again
the bold doesn't "stick" to the text, only when typing within it. This
feels much better than before.

Recorded demo: ![](https://i.imgur.com/gWSLp06.gif)