Fix bug when user pastes non-string
I don’t know how people manage to get non-strings into their clipboard but I was seeing a bunch of errors in sentry "length of undefined" and tracked it down to this plugin.

To re-create on mac

1. Open terminal, execute following command: `pbcopy < /dev/null`
2. Paste in a draft editor that uses this plugin
3. You will see an error in your console

Fix is to early-return if `text` has no value, instead of trying to loop through it.